moved classes out from cqrsintegration tests
