+++ samples/BoltOn.Samples.WebApi/RegistrationTask.cs
@@ -40,10 +40,10 @@ namespace BoltOn.Samples.WebApi
             });
 
 			container.AddTransient<IRepository<Student>, Data.EF.Repository<Student, SchoolDbContext>>();
-			//container.AddTransient<IRepository<Student>, Data.CosmosDb.Repository<Student, SchoolCosmosDbOptions>>();
 			container.AddTransient<IRepository<StudentType>, Data.EF.Repository<StudentType, SchoolDbContext>>();
-			//container.AddTransient<IRepository<StudentFlattened>, Data.CosmosDb.Repository<StudentFlattened, SchoolCosmosDbOptions>>();
 			container.AddTransient<IRepository<StudentFlattened>, Data.EF.Repository<StudentFlattened, SchoolDbContext>>();
+			//container.AddTransient<IRepository<Student>, Data.CosmosDb.Repository<Student, SchoolCosmosDbOptions>>();
+			//container.AddTransient<IRepository<StudentFlattened>, Data.CosmosDb.Repository<StudentFlattened, SchoolCosmosDbOptions>>();
 		}
     }
 }
