+++ tests/BoltOn.Tests/Cqrs/CqrsIntegrationTests.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Extensions.DependencyInjection;
 using Xunit;
@@ -12,10 +11,7 @@ using MassTransit;
 using BoltOn.Mediator.Pipeline;
 using BoltOn.Cqrs;
 using BoltOn.Data.EF;
-using Microsoft.EntityFrameworkCore.Metadata.Builders;
-using Microsoft.EntityFrameworkCore;
 using BoltOn.Data;
-using System.Collections.Generic;
 
 namespace BoltOn.Tests.Cqrs
 {
@@ -341,192 +337,5 @@ namespace BoltOn.Tests.Cqrs
 				.Instance
 				.Dispose();
 		}
-	}
-
-	public class TestCqrsRequest : IRequest
-	{
-		public string Input { get; set; }
-	}
-
-	public class TestCqrsHandler : IRequestAsyncHandler<TestCqrsRequest>,
-		IRequestHandler<TestCqrsRequest>
-	{
-		private readonly IBoltOnLogger<TestCqrsHandler> _logger;
-		private readonly IRepository<TestCqrsWriteEntity> _repository;
-
-		public TestCqrsHandler(IBoltOnLogger<TestCqrsHandler> logger,
-			IRepository<TestCqrsWriteEntity> repository)
-		{
-			_logger = logger;
-			_repository = repository;
-		}
-
-		public void Handle(TestCqrsRequest request)
-		{
-			_logger.Debug($"{nameof(TestCqrsHandler)} invoked");
-			var testCqrsWriteEntity = _repository.GetById(CqrsConstants.EntityId);
-			testCqrsWriteEntity.ChangeInput(request);
-			_repository.Update(testCqrsWriteEntity);
-		}
-
-		public async Task HandleAsync(TestCqrsRequest request, CancellationToken cancellationToken)
-		{
-			_logger.Debug($"{nameof(TestCqrsHandler)} invoked");
-			var testCqrsWriteEntity = await _repository.GetByIdAsync(CqrsConstants.EntityId);
-			testCqrsWriteEntity.ChangeInput(request);
-			await _repository.UpdateAsync(testCqrsWriteEntity, cancellationToken);
-		}
-	}
-
-	public class TestCqrsWriteEntity : BaseCqrsEntity
-	{
-		public string Input { get; internal set; }
-
-		public void ChangeInput(TestCqrsRequest request)
-		{
-			Input = request.Input;
-			RaiseEvent(new TestCqrsUpdatedEvent
-			{
-				Id = CqrsConstants.EventId,
-				Input1 = request.Input,
-				Input2 = new TestInput { Property1 = "prop1", Property2 = 10 }
-			});
-
-			RaiseEvent(new TestCqrsUpdated2Event
-			{
-				Id = CqrsConstants.Event2Id,
-				Input1 = request.Input,
-				Input2 = new TestInput { Property1 = "prop2", Property2 = 20 }
-			});
-		}
-	}
-
-	public class TestCqrsReadEntity : BaseCqrsEntity
-	{
-		public virtual string Input1 { get; internal set; }
-
-		public virtual string Input2Property1 { get; internal set; }
-
-		public virtual int Input2Property2 { get; internal set; }
-
-		public bool UpdateInput(TestCqrsUpdatedEvent @event)
-		{
-			return ProcessEvent(@event, e =>
-			{
-				Input1 = e.Input1;
-				Input2Property1 = e.Input2.Property1;
-				Input2Property2 = e.Input2.Property2;
-			});
-		}
-	}
-
-	public class TestCqrsUpdatedEvent : CqrsEvent
-	{
-		public string Input1 { get; set; }
-
-		public TestInput Input2 { get; set; }
-	}
-
-	public class TestCqrsUpdated2Event : CqrsEvent
-	{
-		public string Input1 { get; set; }
-
-		public TestInput Input2 { get; set; }
-	}
-
-	public class TestInput
-	{
-		public string Property1 { get; set; }
-
-		public int Property2 { get; set; }
-	}
-
-	public class TestCqrsUpdatedEventHandler : IRequestAsyncHandler<TestCqrsUpdatedEvent>,
-		IRequestAsyncHandler<TestCqrsUpdated2Event>
-	{
-		private readonly IBoltOnLogger<TestCqrsUpdatedEventHandler> _logger;
-		private readonly IRepository<TestCqrsReadEntity> _repository;
-
-		public TestCqrsUpdatedEventHandler(IBoltOnLogger<TestCqrsUpdatedEventHandler> logger,
-			IRepository<TestCqrsReadEntity> repository)
-		{
-			_logger = logger;
-			_repository = repository;
-		}
-
-		public async Task HandleAsync(TestCqrsUpdatedEvent request, CancellationToken cancellationToken)
-		{
-			_logger.Debug($"{nameof(TestCqrsUpdatedEventHandler)} invoked");
-			var testCqrsReadEntity = await _repository.GetByIdAsync(request.SourceId);
-			var isSuccessful = testCqrsReadEntity.UpdateInput(request);
-			if (isSuccessful)
-			{
-				_logger.Debug($"{nameof(TestCqrsReadEntity)} updated. " +
-					$"Input1: {testCqrsReadEntity.Input1} Input2Property1: {testCqrsReadEntity.Input2Property1} " +
-					$"Input2Propert2: {testCqrsReadEntity.Input2Property2}");
-				// this is to avoid test warning, as TestCqrsUpdateEvent gets added in two different entities
-				//await _repository.UpdateAsync(testCqrsReadEntity);
-			}
-		}
-
-		public async Task HandleAsync(TestCqrsUpdated2Event request, CancellationToken cancellationToken)
-		{
-			_logger.Debug($"{nameof(TestCqrsUpdatedEventHandler)} invoked for {nameof(TestCqrsUpdated2Event)}");
-			await Task.CompletedTask;
-		}
-	}
-
-	public class TestCqrsWriteEntityMapping : IEntityTypeConfiguration<TestCqrsWriteEntity>
-	{
-		public void Configure(EntityTypeBuilder<TestCqrsWriteEntity> builder)
-		{
-			builder
-				.ToTable("TestCqrsWriteEntity")
-				.HasKey(k => k.Id);
-			builder
-				.HasMany(p => p.EventsToBeProcessed);
-			builder
-				.HasMany(p => p.ProcessedEvents);
-		}
-	}
-
-	public class TestCqrsReadEntityMapping : IEntityTypeConfiguration<TestCqrsReadEntity>
-	{
-		public void Configure(EntityTypeBuilder<TestCqrsReadEntity> builder)
-		{
-			builder
-				.ToTable("TestCqrsReadEntity")
-				.HasKey(k => k.Id);
-			builder
-				.HasMany(p => p.EventsToBeProcessed);
-			builder
-				.HasMany(p => p.ProcessedEvents);
-		}
-	}
-
-	public static class CqrsTestHelper
-	{
-		public static List<string> LoggerStatements { get; set; } = new List<string>();
-	}
-
-	public class CqrsDbContext : BaseDbContext<CqrsDbContext>
-	{
-		public CqrsDbContext(DbContextOptions<CqrsDbContext> options) : base(options)
-		{
-		}
-
-		protected override void ApplyConfigurations(ModelBuilder modelBuilder)
-		{
-			modelBuilder.ApplyConfiguration(new TestCqrsWriteEntityMapping());
-			modelBuilder.ApplyConfiguration(new TestCqrsReadEntityMapping());
-		}
-	}
-
-	public static class CqrsConstants
-	{
-		public static Guid EventId = Guid.Parse("42bc65b2-f8a6-4371-9906-e7641d9ae9cb");
-		public static Guid Event2Id = Guid.Parse("1e5f4d8d-5a3c-478c-a2f5-6bbb43c39a4a");
-		public const string AlreadyProcessedEventId = "90f7f995-c930-4f2c-9621-7c3763a4df1d";
-		public static Guid EntityId = Guid.Parse("b33cac30-5595-4ada-97dd-f5f7c35c0f4c");
 	}
 }
