+++ tests/BoltOn.Tests/Cqrs/CqrsTestHelper.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace BoltOn.Tests.Cqrs
+{
+    public static class CqrsTestHelper
+    {
+        public static List<string> LoggerStatements { get; set; } = new List<string>();
+    }
+}
